Cleaned up TODO's and moved Mover to it's own header
